Extract jsonToZod to a shared function
