+++ packages/agents-core/src/utils/schema-conversion.ts
@@ -14,6 +14,67 @@ type PreviewZodType = z.ZodTypeAny & {
 };
 
 /**
+ * Converts JSON Schema to Zod schema at runtime.
+ *
+ * This is useful for dynamically creating Zod schemas from JSON Schema definitions,
+ * such as when using AI SDK's generateObject/streamObject with data component props.
+ *
+ * Note: Uses z.string() as fallback for unknown types to ensure Anthropic compatibility
+ * (Anthropic requires explicit types - z.any()/z.unknown() produce invalid JSON Schema).
+ */
+export function jsonSchemaToZod(jsonSchema: any): z.ZodType<any> {
+  if (!jsonSchema || typeof jsonSchema !== 'object') {
+    logger.warn({ jsonSchema }, 'Invalid JSON schema provided, using string fallback');
+    return z.string();
+  }
+
+  const schemaType = jsonSchema.type as string | undefined;
+
+  switch (schemaType) {
+    case 'object': {
+      const properties = jsonSchema.properties as
+        | Record<string, Record<string, unknown>>
+        | undefined;
+      if (properties && typeof properties === 'object') {
+        const shape: Record<string, z.ZodType<unknown>> = {};
+        for (const [key, prop] of Object.entries(properties)) {
+          shape[key] = jsonSchemaToZod(prop);
+        }
+        return z.object(shape);
+      }
+      // Object without defined properties - use record with string values as safe fallback
+      return z.record(z.string(), z.string());
+    }
+
+    case 'array': {
+      const items = jsonSchema.items as Record<string, unknown> | undefined;
+      const itemSchema = items ? jsonSchemaToZod(items) : z.string();
+      return z.array(itemSchema);
+    }
+
+    case 'string':
+      return z.string();
+
+    case 'number':
+    case 'integer':
+      return z.number();
+
+    case 'boolean':
+      return z.boolean();
+
+    case 'null':
+      return z.null();
+
+    default:
+      logger.warn(
+        { unsupportedType: schemaType, schema: jsonSchema },
+        'Unsupported JSON schema type, using string fallback'
+      );
+      return z.string();
+  }
+}
+
+/**
  * Utility function for converting Zod schemas to JSON Schema
  * Uses Zod's built-in toJSONSchema method
  */
